Rocks Layout: Fix a bug in implementation of `leading`
It was possible to have a situation like the following:
          ........
          .      .
      ____.  __  .
     |____| |__| .
~~~~~~~~~~.~~~~~~.~~~~~~~~~
          ........

where ~~~~ stands in for the chunk boundary, and ... is the maximum
padding of the 2nd fragment. If we were testing some fragment against
the fragments in the `backing`, then we would find a collision in the
bottom-most chunk and exit. However, this behavior isn't correct;
suppose that the fragment we're testing against is _compatible_ with
the 2nd fragment, but incompatible with the first. Then, it's quite
possible that a larger offset will derive from the 1st fragment, but
we'll never check it since we stop after scanning the bottom most
chunk.

So, instead, we implement a rule where if the fragment we're testing
against, with its maximal padding applied, crosses the chunk boundary
when placed at the current maximum offset, then we'll check the above
chunk as well.