Rocks Layout: Improve performance of the Timetable
Instead of a cell being a pair of padding and UID, we allow cells to
additionally be a reference to (i.e. the index of) another cell in
their column. Then, when building the timetable, we add a reference
when we `fillColumn`, rather than adding a bunch of identical cells.

This change means that when we find the `spaceBetween` two stacks, we
can immediately skip to a relevant cell in the table, rather than
iterating through many (potentially identical) cells.

Care has also been taken to inline array access, rather than deferring
to helpers such as `getCell`, which was found to be very slow through
profiling (in no small part due to its assertion statements). It's too
bad to have to remove these assertions, but I think the extra
performance is worth it.

This change makes the `core.ts` benchmark take 0.27 seconds rather
than 4.1 seconds.