Rocks + Pebble: Implement algorithm G1
This mainly requires changing the definition of `RegionWithLeading` to
keep track of its origin as well as its advance, and piping this
change through the requisite functions.

Algorithm G1 has been exposed not as a completely separate algorithm,
but as a flag which can be adjusted on both L1S + L1S+, and
L1P. Additional layout tests have been added to rock and pebble which
exercise this new flag.

This change can be seen as adding a "lead in" point in addition to the
implicit "lead-out" point that we already had. In the previous version
of the algorithm, we assumed that the "lead-in" point was always at
the origin. This is why when we wrapped regions, we had to translate
them -- the lead-in point _must_ remain at (0, 0) so that when we
place regions next-to each other, the wraps are accounted for.

However, this choice has a consequence: when we merge the region
corresponding to each line into a single region at the end of layout,
we don't (can't) "look inside" them, and so concatenate them so that
their origins line up on a vertical line. Since we've translated the
contents of the region, though, the first fragment on a line might not
be at the origin.

But now, since we remember how much we've translated the first
fragment of a region (the region's _origin_), we don't need to
translate the region when we wrap -- just adjust the origin. So, the
first fragment of a line stays in place, resulting in more compact
layouts.