Demo: Update example programs
Now, all frames of the "CLOC" example from the CSS paper are
represented as built-in examples, although not all of them are
completely styled. The trace example can't be rendered well since
there's no way to place an arbitrary box, and the final frame can't be
rendered well since we cannot yet style the color of text.